pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "link"
version = "0.1.0"
description = "智能工具中间件 - 通过 Matrix 协议连接用户与工具的 Agent 框架"
requires-python = ">=3.11"
dependencies = [
    "matrix-nio>=0.24.0",
    "openai>=1.0.0",
    "aiohttp>=3.9.0",
    "aiofiles>=23.0.0",
    "pyyaml>=6.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
ltool = "link.main:main"

[tool.setuptools.packages.find]
include = ["link*"]
